backend/requirements.txt
# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# python-multipart==0.0.6
# pydantic==2.5.3
# motor==3.3.2
# pymongo==4.6.1
# joblib==1.3.2
# transformers==4.37.2
# sentence-transformers==2.3.1
# safetensors==0.4.1
# sentencepiece==0.1.99
# Pillow==10.2.0
# numpy==1.26.3
# python-dotenv==1.0.0
# requests==2.31.0
# gunicorn==21.2.0
# aiofiles==23.2.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
aiofiles==23.2.1

# Gemini SDK
google-genai>=0.3.0
motor==3.3.2
bcrypt==4.1.2
padding-oracle==1.0
passlib[bcrypt]==1.7.4